gain fix, defaults changed, and minor refactor

+++ src/hardware/pslab/protocol.h
@@ -52,7 +52,7 @@
 #define SET_LO_CAPTURE 0x10
 #define RETRIEVE_BUFFER 0x08
 
-static const uint8_t GAIN_VALUES[] = {1, 2, 4, 5, 8, 10, 16, 32};
+static const uint64_t GAIN_VALUES[] = {1, 2, 4, 5, 8, 10, 16, 32};
 
 struct dev_context {
 	/* device mode */
